requirements.txt
asgiref==3.4.1
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.9
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
daphne==3.0.2
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==1.0.0
Django==3.2.8
django-ckeditor==6.1.0
django-cors-headers==3.10.0
django-crispy-forms==1.13.0
django-decouple==2.1
django-email-verification==0.1.0
django-environ==0.8.1
django-filter==21.1
django-guardian==2.4.0
django-js-asset==1.2.2
django-phonenumber-field==6.1.0
django-rest-knox==4.1.0
django-rest-paystack==2.2.2
django-tables2==2.4.1
django-templated-mail==1.1.1
django-tinymce==3.3.0
django-tinymce4==3.0.2
Django-Verify-Email==1.0.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.4
filelock==3.4.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.4
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.3
phonenumbers==8.12.48
Pillow==9.2.0
platformdirs==2.4.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.10.0
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-decouple==3.6
python-imap==1.0.0
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
reportlab==3.6.11
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
secure-smtplib==0.1.1
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
stripe==2.63.0
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.10.0
whitenoise==6.2.0
zope.interface==5.4.0
